feat: Created 'FormLogin' and scss files
