+++ frontend/src/app/Login/page.tsx
@@ -1,35 +1,9 @@
 "use client";
-/* eslint-disable @typescript-eslint/no-unused-vars */
-/* eslint-disable @typescript-eslint/no-unused-expressions */
 
-import usePost from "@/services/usePost";
-import Link from "next/link";
-import React, { useState } from "react";
-
-interface ILogin {
-  email: string;
-  password: string;
-}
+import { FormLogin } from "@/components/Forms/Login/FormLogin";
+import "./style.scss";
 
 export default function Login() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const { cadastrarDados, error, sucesso } = usePost();
-
-  const handleLogin = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    const usuario: ILogin = {
-      email: email,
-      password: password,
-    };
-
-    try {
-      cadastrarDados({ url: "auth/login", dados: usuario });
-    } catch (error) {
-      error && alert("Não foi possível realizar o login!");
-    }
-  };
-
   return (
     <div className="container d-flex flex-column flex-grow-1 text-black">
       <div className="container">
@@ -40,58 +14,20 @@ export default function Login() {
         <hr className="border border-warning border-2"></hr>
         <h2 className="d-flex justify-content-center mb-4">Acesse sua conta</h2>
       </div>
-      <div className="container d-flex justify-content-center mb-5">
-        <form className="row g-3" onSubmit={handleLogin}>
-          <div className="col-12 mb-3">
-            <label className="form-label">
-              E-mail
-              <span className="text-danger ms-1">*</span>
-            </label>
-            <input
-              type="email"
-              className="form-control"
-              id="email"
-              placeholder="exemplo@ufba.br"
-              required
-            />
-          </div>
-          <div className="col-12 mb-3">
-            <label className="form-label">
-              Senha
-              <span className="text-danger ms-1">*</span>
-            </label>
-            <input
-              type="password"
-              className="form-control"
-              id="password"
-              placeholder="digite sua senha"
-              required
-            />
 
-            <div className="text-end">
-              <a href="/AlterarSenha" className="link-underline link-underline-opacity-0" style={{ color: "blue" }}>
-                Esqueceu sua senha
-              </a>
-            </div>
-          </div>
-          <div className="d-grid gap-2 col-3 mx-auto">
-            <button type="submit" className="btn btn-primary">
-              Entrar
-            </button>
-          </div>
-          <hr className="border border-warning border-2"></hr>
-        </form>
+      <div className="container d-flex justify-content-center mb-5">
+        <FormLogin />
       </div>
+
       <div className="text-center mb-4">
         <h6>
           Ainda não tem conta?
-          <Link
+          <a
             href="/Cadastro"
             className="link-underline link-underline-opacity-0 ms-1"
-            style={{ color: "blue" }}
           >
             Cadastre-se
-          </Link>
+          </a>
         </h6>
       </div>
     </div>
