+++ frontend/src/app/Login/style.scss
@@ -1,0 +1,3 @@
+a {
+    color: blue;
+}
